Add RequireJS
