+++ admin/src/main/resources/io/buoyant/admin/.eslintrc.json
@@ -3,11 +3,14 @@
     "browser": true,
     "jquery": true
   },
-  "extends": "eslint:recommended",
+  "extends": ["eslint:recommended", "plugin:requirejs/recommended"],
   "globals": {
     "_": true,
     "Handlebars": true
   },
+  "plugins": [
+    "eslint-plugin-requirejs"
+  ],
   "rules": {
     "no-console": ["warn", { "allow": ["warn", "error"] }],
     "no-unused-vars": ["error", { "argsIgnorePattern": "^_", "varsIgnorePattern": "^_" }]
