+++ admin/src/main/resources/io/buoyant/admin/js/app/metrics.js
