--- admin/src/main/resources/io/buoyant/admin/spec/helpers/jasmine_examples/SpecHelper.js
@@ -1,15 +1,0 @@
-beforeEach(function () {
-  jasmine.addMatchers({
-    toBePlaying: function () {
-      return {
-        compare: function (actual, expected) {
-          var player = actual;
-
-          return {
-            pass: player.currentlyPlayingSong === expected && player.isPlaying
-          }
-        }
-      };
-    }
-  });
-});
